refactor: page->pages, store->stores 디렉토리명 변경
