+++ src/components/DashboardData.vue
@@ -74,7 +74,7 @@
 </template>
 
 <script>
-import { useWebSocketStore } from "@/store";
+import { useWebSocketStore } from "@/stores";
 import { ref, watch, reactive, onMounted, onUnmounted } from "vue";
 import Chart from "primevue/chart";
