+++ src/pages/CapturePage.vue
