+++ src/stores/websocket/types.ts
